A lesson to help beginners dodge various syntax traps
With exercises that may be a little too frightening...
Should we have this so early? Should we have it at all?
I'm under the impression that ocaml syntax is too complex
to allow a complete beginner to take advantage of the current
tutorial, yet I'm not convinced this is a good idea to draw
attention on this topic.